Official Configuration API for addons
The one class limit for addons (it may be possible to have more, but
it's a mess) causes a problem. If an addon wants a config they need to
have a mess of code simply because theres no config file bundled in the
jar for bukkit to extract. This can be solved however by installing an
config from an URL. All done natively, safely, and best of all SIMPLY!